Add ssh-perms target to GNUmakefile

+++ GNUmakefile
@@ -9,7 +9,7 @@ HOSTNAME = `uname -n`
 # command to set the nodump attribute
 NODUMP = chattr +d 
 
-all: rc_files dropbox_files insync_files no_dump vim-helptags
+all: rc_files dropbox_files insync_files no_dump ssh-perms vim-helptags
 
 rc_files: .vimrc .gvimrc .irbrc
 
@@ -18,12 +18,8 @@ dropbox_files:
 	if [ -d ~/Dropbox ]; then chmod -R u=rwX,g=rX,o=  ~/Dropbox; else true; fi
 	if [ -d ~/Dropbox/Backups -a ! -d ~/Backups ]; then ln -s ~/Dropbox/Backups ~/Backups; else true; fi
 	if [ -d ~/Dropbox/Documents -a ! -d ~/Documents ]; then ln -s ~/Dropbox/Documents ~/Documents; else true; fi
-	if [ ! -h ~/.ssh/keys -a ! -e ~/.ssh/keys ]; then \
-		ln -s ~/Dropbox/Ssh/keys ~/.ssh/keys; \
-		find ~/.ssh/keys/ -type d -exec chmod 0700 '{}' \; ; \
-		find ~/.ssh/keys/ -type f -exec chmod 0600 '{}' \; ; fi
-	if [ ! -h .ssh/config -a ! -f .ssh/config ]; then \
-		ln -s ~/Dropbox/Ssh/config ~/.ssh/config; fi
+	if [ ! -h ~/.ssh/keys -a ! -e ~/.ssh/keys ]; then ln -s ~/Dropbox/Ssh/keys ~/.ssh/keys; fi
+	if [ ! -h .ssh/config -a ! -f .ssh/config ]; then ln -s ~/Dropbox/Ssh/config ~/.ssh/config; fi
 
 insync_files:
 	if [ -d ~/bigboss1964@gmail.com ]; then chmod -R u=rwX,g=rX,o=  ~/bigboss1964@gmail.com; else true; fi
@@ -48,6 +44,12 @@ push:
 .irbrc:
 	touch .irbrc
 
+ssh-perms:
+		chmod -R u=rwX,g=,o= ~/.ssh/keys
+		chmod 0700 ~/.ssh/config
+		chmod 0700 ~/.ssh/authorized_keys
+		chmod 0700 ~/.ssh/known_hosts
+
 vim-helptags:
 	if type vim >/dev/null 2>/dev/null; then \
 		for _D in .vim/bundle/*/doc; do \
@@ -62,5 +64,5 @@ no_dump:
 	if [ -d ~/.dbus/session-bus ]; then $(NODUMP) ~/.dbus/session-bus ; else true; fi
 	if [ -d ~/.m2/repository ]; then $(NODUMP) ~/.m2/repository ; else true; fi
 
-.PHONY: status pull push dropbox_files insync_files vim-helptags no_dump
+.PHONY: status pull push dropbox_files insync_files ssh-perms vim-helptags no_dump
